+++ src/Model/Export/Category/Field/ParentCategory.php
@@ -6,16 +6,25 @@ namespace Omikron\Factfinder\Model\Export\Category\Field;
 
 use Magento\Framework\Model\AbstractModel;
 use Omikron\Factfinder\Api\Export\FieldInterface;
+use Omikron\Factfinder\Model\Formatter\CategoryPathFormatter;
 
 class ParentCategory implements FieldInterface
 {
+    /** @var CategoryPathFormatter  */
+    private $categoryPathFormatter;
+
+    public function __construct(CategoryPathFormatter $categoryPathFormatter)
+    {
+        $this->categoryPathFormatter = $categoryPathFormatter;
+    }
+
     public function getName(): string
     {
-        return 'ParentCategory';
+        return 'parentCategory';
     }
 
     public function getValue(AbstractModel $category): string
     {
-        return $category->getPathInStore();
+        return $this->categoryPathFormatter->format((int) $category->getParentId(), $category->getStore());
     }
 }